Expanded development details in the README
Noticed that `./gradlew build` does not always rebuild the
API models based on the latest code changes, resulting in
the locally generated build artifacts remaining unchanged
and hiding build errors that will show up in deployed
environments.

This is because Gradle locally caches build artifacts, and
in order to force a rebuild based on the latest local code,
we must run `./gradlew clean build` to clean and rebuild the
project after having successfully run `./gradlew build` at
least once.

Updated the README with additional details on building
and testing changes.