Added script to generate ReDoc API documentation
ReDoc is an open-source tool for generating API documentation
from OpenAPI specs.

In order to generate ReDoc from our Smithy-generated OpenAPI
JSON spec, we must convert the JSON spec to YAML, the format
required by ReDoc, and then run the redoc-cli with the
YAML spec as an input.

Added a scripts/generate-docs.sh to automate this process
with the prerequisite that the yq tool for converting
JSON to YAML, and redoc-cli, have been pre-installed.