Add OpenAPI JSON spec to source control
Add the generated OpenAPI JSON spec to source control
to make it easier for maintainers to see the impact
of model changes on the spec, and make it easier to
integrate with downstream workflows that depend on
the OpenAPI spec.

Prior to this change, downstream workflows depending
on the spec would need to have all of this package's
dependencies installed and gradle build the project
to generate the spec.

Add Gradle task to copy the generated OpenAPI
JSON spec to openapi/ConsentManagementApi.openapi.json
after every build.